+++ src/icalendar/tests/test_icalendar.py
@@ -215,7 +215,9 @@ class IcalendarTestCase(unittest.TestCase):
         with pytest.raises(
             ValueError, match="Content line could not be parsed into parts"
         ):
-            Contentline("REFRESH - INTERVAL; VALUE = DURATION:PT48H", strict=True).parts()
+            Contentline(
+                "REFRESH - INTERVAL; VALUE = DURATION:PT48H", strict=True
+            ).parts()
 
         contains_base64 = (
             b"X-APPLE-STRUCTURED-LOCATION;"